fix : entity 하나로 합침
