+++ src/main/java/nbc/devmountain/domain/lecture/model/LectureCategoryMapping.java
@@ -4,6 +4,8 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -11,22 +13,26 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "lecture_category")
+@Table(name = "lecture_category_mapping")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class LectureCategory {
+public class LectureCategoryMapping {
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long lectureCategoryId;
-	private int id;
-	private String title;
 
+	@ManyToOne
+	@JoinColumn(name = "lecture_category_id")
+	private LectureCategory lectureCategory;
 
+	@ManyToOne
+	@JoinColumn(name = "lecture_id")
+	private Lecture lecture;
 
 	@Builder
-	public LectureCategory(int id, String title){
-		this.id = id;
-		this.title = title;
+	public LectureCategoryMapping(LectureCategory lectureCategory, Lecture lecture){
+		this.lectureCategory = lectureCategory;
+		this.lecture = lecture;
 	}
-
 }
