+++ src/main/java/nbc/devmountain/domain/lecture/service/LectureService.java
@@ -1,13 +1,19 @@
 package nbc.devmountain.domain.lecture.service;
 
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
 import nbc.devmountain.domain.lecture.client.LectureClient;
+import nbc.devmountain.domain.lecture.dto.CategoryDto;
 import nbc.devmountain.domain.lecture.dto.InflearnResponse;
 import nbc.devmountain.domain.lecture.dto.Item;
+import nbc.devmountain.domain.lecture.model.LectureCategory;
 import nbc.devmountain.domain.lecture.model.Lecture;
-import nbc.devmountain.domain.lecture.model.LectureMapper;
+import nbc.devmountain.domain.lecture.model.LectureCategoryMapping;
+import nbc.devmountain.domain.lecture.repository.LectureCategoryRepository;
+import nbc.devmountain.domain.lecture.repository.LectureCategoryMappingRepository;
 import nbc.devmountain.domain.lecture.repository.LectureRepository;
 
 @Service
@@ -16,7 +22,8 @@ public class LectureService {
 
 	private final LectureClient lectureClient;
 	private final LectureRepository lectureRepository;
-	private final LectureMapper lectureMapper;
+	private final LectureCategoryRepository lectureCategoryRepository;
+	private final LectureCategoryMappingRepository lectureCategoryMappingRepository;
 
 	public void getLecture() {
 
@@ -38,14 +45,46 @@ public class LectureService {
 	private void savePage(InflearnResponse page) {
 
 		for (Item item : page.data().items()) {
+			List<LectureCategory> lectureCategoryList = lectureCategoryRepository.saveAll(
+				toCategoryEntity(item.course().metadata().categories()));
 
-		Lecture lecture	= lectureMapper.toEntity(item);
+			Lecture lecture = lectureRepository.save(Lecture.builder()
+				.itemId(item.id())
+				.thumbnailUrl(item.course().thumbnailUrl())
+				.title(item.course().title())
+				.instructor(item.instructor().name())
+				.description(item.course().description())
+				.reviewCount(item.course().reviewCount())
+				.studentCount(item.course().studentCount())
+				.likeCount(item.course().likeCount())
+				.star(item.course().star())
+				.levelCode(item.course().metadata().levelCode())
+				.isDiscount(item.listPrice().isDiscount())
+				.payPrice(item.listPrice().payPrice())
+				.regularPrice(item.listPrice().regularPrice())
+				.isFree(item.listPrice().isFree())
+				.discountRate(item.listPrice().discountRate())
+				.build());
 
-		lectureRepository.save(lecture);
+			lectureCategoryMappingRepository.saveAll(lectureCategoryList.stream()
+				.map(c -> LectureCategoryMapping.builder()
+					.lecture(lecture)
+					.lectureCategory(c)
+					.build())
+				.toList());
 
 		}
-
 	}
+
+	private List<LectureCategory> toCategoryEntity(List<CategoryDto> categoryDto) {
+		return categoryDto.stream()
+			.map(c -> LectureCategory.builder()
+				.id(c.id())
+				.title(c.title())
+				.build())
+			.toList();
+	}
+
 }
